requirements.txt
fastapi
uvicorn
streamlit
pandas
plotly
sqlalchemy
langgraph
langchain-groq
langchain-huggingface
mem0ai
apscheduler
python-dotenv
requests
sentence-transformers
qdrant-client
langchain-openai
